+++ tendrl/commons/objects/definition/master.yaml
@@ -24,6 +24,8 @@ namespace.tendrl:
     CreateCluster:
       tags: 
         - "provisioner/$TendrlContext.sds_name"
+      atoms:
+        - tendrl.objects.Cluster.atoms.Create
       help: "Create a cluster from scratch"
       enabled: true
       inputs:
@@ -40,6 +42,11 @@ namespace.tendrl:
           - Cluster.conf_overrides
       pre_run:
         - tendrl.objects.Node.atoms.IsNodeTendrlManaged
+        - tendrl.objects.Cluster.atoms.ValidCreateClusterParams
+      post_run:
+        - tendrl.objects.Cluster.atoms.PersistClusterDetailsAfterCreate
+        - tendrl.objects.Cluster.atoms.ImportCreatedCluster
+        - tendrl.objects.Cluster.atoms.CheckClusterAvailable
       run: tendrl.flows.CreateCluster
       type: Create
       uuid: 2f94a48a-05d7-408c-b400-e27827f4eacd
@@ -68,6 +75,8 @@ namespace.tendrl:
     ImportCluster:
       tags:
         - "detected_cluster/$DetectedCluster.detected_cluster_id"
+      atoms:
+        - tendrl.objects.Cluster.atoms.ImportCluster
       help: "Import existing Gluster Cluster"
       enabled: true
       inputs:
@@ -77,6 +86,10 @@ namespace.tendrl:
           - TendrlContext.integration_id
       pre_run:
         - tendrl.objects.Node.atoms.IsNodeTendrlManaged
+        - tendrl.objects.Cluster.atoms.ValidImportClusterParams
+        - tendrl.objects.Cluster.atoms.CheckNodesInExistingCluster
+      post_run:
+        - tendrl.objects.Cluster.atoms.CheckClusterAvailable
       run: tendrl.flows.ImportCluster
       type: Create
       uuid: 2f94a48a-05d7-408c-b400-e27827f4edef
@@ -88,6 +101,129 @@ namespace.tendrl:
       help: "Represents a cluster"
       list: /clusters
       value: clusters/$TendrlContext.integration_id
+      atoms:
+        ValidCreateClusterParams:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - TendrlContext.cluster_name
+              - TendrlContext.sds_name
+              - TendrlContext.cluster_id
+              - TendrlContext.integration_id
+              - Cluster.node_configuration
+            optional:
+              - Cluster.public_network
+              - Cluster.cluster_network
+              - Cluster.conf_overrides
+          name: valid create cluster params
+          run: tendrl.objects.Cluster.atoms.ValidCreateClusterParams
+          type: Get
+          uuid: 352f6190-9b37-11e6-950d-a24fc0d9649d
+          help: Check if the create cluster parameters are valid
+        ValidImportClusterParams:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - DetectedCluster.sds_pkg_name
+              - TendrlContext.integration_id
+          name: valid import cluster params
+          run: tendrl.objects.Cluster.atoms.ValidImportClusterParams
+          type: Get
+          uuid: 352f6190-9b37-11e6-950d-a24fc0d9649e
+          help: Check if the import cluster parameters are valid
+        ImportCreatedCluster:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - TendrlContext.cluster_name
+              - TendrlContext.sds_name
+              - TendrlContext.cluster_id
+              - TendrlContext.integration_id
+              - Cluster.node_configuration
+            optional:
+              - Cluster.public_network
+              - Cluster.cluster_network
+              - Cluster.conf_overrides
+          name: import created cluster
+          run: tendrl.objects.Cluster.atoms.ImportCreatedCluster
+          type: Update
+          uuid: 352f6190-9b37-11e6-950d-a24fc0d9649f
+          help: Import the created cluster
+        PersistClusterDetailsAfterCreate:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - TendrlContext.cluster_name
+              - TendrlContext.sds_name
+              - TendrlContext.cluster_id
+              - TendrlContext.integration_id
+              - Cluster.node_configuration
+            optional:
+              - Cluster.public_network
+              - Cluster.cluster_network
+              - Cluster.conf_overrides
+          name: persist cluster additional details
+          run: tendrl.objects.Cluster.atoms.PersistClusterDetailsAfterCreate
+          type: Update
+          uuid: 352f6190-9b37-11e6-950d-a24fc0d9649a
+          help: Persist additional cluster details
+        Create:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - TendrlContext.cluster_name
+              - TendrlContext.sds_name
+              - TendrlContext.cluster_id
+              - TendrlContext.integration_id
+              - Cluster.node_configuration
+            optional:
+              - Cluster.public_network
+              - Cluster.cluster_network
+              - Cluster.conf_overrides
+          name: create cluster
+          run: tendrl.objects.Cluster.atoms.Create
+          type: Create
+          uuid: 352f6190-9b37-11e6-950d-a24fc0d9649b
+          help: create cluster
+        ImportCluster:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - DetectedCluster.sds_pkg_name
+              - TendrlContext.integration_id
+          name: import cluster
+          run: tendrl.objects.Cluster.atoms.Import
+          type: Create
+          uuid: 452f6190-9b37-11e6-950d-a24fc0d9649c
+          help: Import the cluster
+        CheckClusterAvailable:
+          enabled: true
+          inputs:
+            mandatory:
+              - TendrlContext.integration_id
+          name: check cluster available
+          run: tendrl.objects.Cluster.atoms.CheckClusterAvailable
+          type: Get
+          uuid: 552f6190-9b37-11e6-950d-a24fc0d9649e
+          help: check if cluster is available
+        CheckNodesInExistingCluster:
+          enabled: true
+          inputs:
+            mandatory:
+              - "Node[]"
+              - DetectedCluster.sds_pkg_name
+              - TendrlContext.integration_id
+          name: check nodes in existing cluster
+          run: tendrl.objects.Cluster.atoms.CheckNodesInExistingCluster
+          type: Get
+          uuid: 652f6190-9b37-11e6-950d-a24fc0d9649c
+          help: node is not part of existing cluster
       attrs:
         integration_id:
           help: integration id
