+++ tendrl/commons/objects/tendrl_context/__init__.py
@@ -4,6 +4,7 @@ from tendrl.commons import objects
 class TendrlContext(objects.BaseObject):
     def __init__(
         self,
+        node_id=None,
         integration_id=None,
         cluster_id=None,
         cluster_name=None,
@@ -11,6 +12,7 @@ class TendrlContext(objects.BaseObject):
         sds_version=None,
             *args, **kwargs):
         super(TendrlContext, self).__init__(*args, **kwargs)
+        self.node_id = node_id
         # integration_id is the Tendrl generated cluster UUID
         self.integration_id = integration_id
         self.cluster_id = cluster_id
@@ -20,5 +22,5 @@ class TendrlContext(objects.BaseObject):
         self.value = 'nodes/{0}/TendrlContext'
 
     def render(self):
-        self.value = self.value.format(NS.node_context.node_id)
+        self.value = self.value.format(node_id or NS.node_context.node_id)
         return super(TendrlContext, self).render()
